Update README.md
All movies and series listed in my favorites are basically horror films. It just gives me the thrill specially watching it in the middle of the night. There are some unexpected plots that gives you a "guessing game" vibe. Also finding out the villain and reason of the plot is quite thrilling.